requirements.txt
crewai>=0.50.0
crewai-tools>=0.10.0
fastapi>=0.110.0
uvicorn>=0.27.0
yfinance>=0.2.36
sentence-transformers>=2.5.1
litellm>=1.30.0
streamlit>=1.31.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pytest>=8.0.0
requests>=2.31.0
